requirements.txt
ipython==8.2.0
matplotlib==3.5.1
numpy==1.22.2
pandas==1.4.0
Pillow==9.0.1
tensorflow==2.8.0
tensorflow_addons==0.15.0
tensorflow_gpu==2.8.0
tensorflow_io==0.24.0
osgeo==3.2.3
cryptography==3.4.7
arcgis==1.8.5.post3
